Automation/root/requirements.txt
pytest==8.4.0
pytest-html==4.1.1
requests==2.32.4
selenium==4.33.0
allure-pytest